Module html2text::render::text_renderer

Implementations of the Renderer trait.

This module implements helpers and concrete types for rendering from HTML into different text formats.

Structs

BorderHoriz
A dividing line between table rows which tracks intersections with vertical lines.

PlainDecorator
A decorator for use with TextRenderer which outputs plain UTF-8 text with no annotations. Markup is rendered as text characters or footnotes.

RichDecorator
A decorator to generate rich text (styled) rather than pure text output.

TaggedLine
A line of tagged text (composed of a set of TaggedStrings).

TaggedString
A wrapper around a String with extra metadata.

TextRenderer
A renderer which just outputs plain text with annotations depending on a decorator.

TrivialDecorator
A decorator for use with TextRenderer which outputs plain UTF-8 text with no annotations or markup, emitting only the literal text.

Enums

BorderSegHoriz
A space on a horizontal row.

RenderLine
A line, which can either be text or a line.

RichAnnotation
Annotation type for "rich" text. Text is associated with a set of these.

TaggedLineElement
An element of a line of tagged text: either a TaggedString or a marker appearing in between document characters.

Traits

TextDecorator
Allow decorating/styling text.