Struct html2text::render::text_renderer::TaggedLine

pub struct TaggedLine<T: Debug + Eq + PartialEq + Clone> { /* fields omitted */ }

A line of tagged text (composed of a set of TaggedStrings).

Methods

impl<T: Debug + Eq + PartialEq + Clone + Default> TaggedLine<T>

pub fn new() -> TaggedLine<T>

Create an empty TaggedLine.

pub fn from_string(s: String, tag: &T) -> TaggedLine<T>

Create a new TaggedLine from a string and tag.

pub fn into_string(self) -> String

Join the line into a String, ignoring the tags and markers.

pub fn is_empty(&self) -> bool

Return true if the line is non-empty

pub fn push_str(&mut self, ts: TaggedString<T>)

Add a new tagged string fragment to the line

pub fn push(&mut self, tle: TaggedLineElement<T>)

Add a new general TaggedLineElement to the line

pub fn insert_front(&mut self, ts: TaggedString<T>)

Add a new fragment to the start of the line

pub fn push_char(&mut self, c: char, tag: &T)

Add text with a particular tag to self

pub fn consume(&mut self, tl: &mut TaggedLine<T>)

Drain tl and use to extend self.

pub fn drain_all(&mut self) -> Drain<TaggedLineElement<T>>

Drain the contained items

pub fn chars<'a>(&'a self) -> Box<dyn Iterator<Item = char> + 'a>

Iterator over the chars in this line.

pub fn iter<'a>(
&'a self
) -> Box<dyn Iterator<Item = &TaggedLineElement<T>> + 'a>

Iterator over TaggedLineElements

pub fn width(&self) -> usize

Return the width of the line in cells

pub fn pad_to(&mut self, width: usize)

Pad this line to width with spaces (or if already at least this wide, do nothing).