Struct html2text::render::text_renderer::TextRenderer

pub struct TextRenderer<D: TextDecorator> { /* fields omitted */ }

A renderer which just outputs plain text with annotations depending on a decorator.

Methods

impl<D: TextDecorator> TextRenderer<D>

pub fn new(width: usize, decorator: D) -> TextRenderer<D>

Construct a new empty TextRenderer.

pub fn add_subblock(&mut self, s: &str)

Add a prerendered (multiline) string with the current annotations.

pub fn into_string(self) -> String

Consumes this renderer and return a multiline String with the result.

pub fn into_lines(self) -> Vec<RenderLine<Vec<D::Annotation>>>

Returns a Vec of TaggedLines with therendered text.